refactor(examples): 迁移 scroller 至 plugins 目录
