+++ examples/src/pages/plugins/scroller/index.less
