+++ examples/src/pages/plugins/scroller/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Button } from 'antd'
 import { Graph, Export, Selection, MiniMap, Scroller } from '@antv/x6'
-import '../index.less'
+import '../../index.less'
 import './index.less'
 
 export class ScrollerExample extends React.Component {
